fix: stabilize machine state and cleanroom tracking

+++ src/main/java/io/github/meistermods/siliconic/client/CoalGeneratorScreen.java
@@ -25,7 +25,10 @@ public class CoalGeneratorScreen extends AbstractContainerScreen<CoalGeneratorMe
         slotBox(g, leftPos + 8 + column * 18, topPos + 96 + row * 18);
     for (int column = 0; column < 9; column++) slotBox(g, leftPos + 8 + column * 18, topPos + 154);
 
-    int flameHeight = menu.totalBurnTime() == 0 ? 0 : 16 * menu.burnTime() / menu.totalBurnTime();
+    int flameHeight =
+        menu.totalBurnTime() == 0
+            ? 0
+            : (int) (16L * menu.burnTime() / menu.totalBurnTime());
     g.fill(leftPos + 66, topPos + 38, leftPos + 78, topPos + 54, 0xff2b3035);
     g.fill(leftPos + 66, topPos + 54 - flameHeight, leftPos + 78, topPos + 54, 0xffff9f43);
     int energyWidth = menu.capacity() == 0 ? 0 : 160 * menu.energy() / menu.capacity();
